backend/pyproject.toml
[project]
name = "rpa-api"
version = "0.1.0"
description = "FastAPI service for RPA automation"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "playwright",
    "pydantic-settings",
    "python-dotenv",
    "requests",
    "uvicorn",
]

[tool.uv]
package = false
